test(radio): recategorize and clean up browser tests
